requirements.txt
asgiref==3.11.0
dj-database-url==2.0.0
Django==4.2.10
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.5.5
whitenoise==6.6.0
